+++ common/bolt11.c
@@ -467,6 +467,14 @@ struct bolt11 *bolt11_decode(const tal_t *ctx, const char *str,
 
         b11->routes = tal_arr(b11, struct route_info *, 0);
 
+	/* BOLT #11:
+	 *
+	 * If a URI scheme is desired, the current recommendation is to either
+	 * use 'lightning:' as a prefix before the BOLT-11 encoding
+	 */
+	if (strstarts(str, "lightning:") || strstarts(str, "LIGHTNING:"))
+		str += strlen("lightning:");
+
         if (strlen(str) < 8)
                 return decode_fail(b11, fail, "Bad bech32 string");